Reconfigure file structure for database growth
