+++ .env
@@ -1,1 +1,2 @@
 PORT=4000
+JWT_SECRET=There is no secret.  You had the power the whole time.
