+++ auth/restrictedMiddleware.js
@@ -1,0 +1,20 @@
+const jwt = require("jsonwebtoken");
+
+const secrets = require("../config/secrets.js");
+
+module.exports = (req, res, next) => {
+  const token = req.headers.authorization;
+
+  if (token) {
+    jwt.verify(token, secrets.jwtSecret, (error, decodedToken) => {
+      if (error) {
+        res.status(401).json({ message: "Can't touch this!" });
+      } else {
+        req.decodedJwt = decodedToken;
+        next();
+      }
+    });
+  } else {
+    res.status(401).json({ message: "YOU SHALL NOT PASS!" });
+  }
+};
