+++ config/secrets.js
@@ -1,0 +1,3 @@
+module.exports = {
+  jwtSecret: process.env.JWT_SECRET || "The world is quiet here."
+};
