+++ users/usersRouter.js
@@ -1,80 +1,11 @@
-const express = require("express");
-const bcrypt = require("bcryptjs");
-const jwt = require("jsonwebtoken");
+const router = require("express").Router();
 
 const Users = require("./usersModel.js");
+const restricted = require("../auth/restrictedMiddleware.js");
 
-const router = express.Router();
-
-const secret = "The world is quiet here.";
-
-router.post("/register", (req, res) => {
-  let user = req.body;
-  const hash = bcrypt.hashSync(user.password, 10);
-  user.password = hash;
-
-  Users.add(user)
-    .then(saved => {
-      res.status(201).json(saved);
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-function generateToken(user) {
-  const payload = {
-    subject: user.id,
-    username: user.username
-  };
-
-  const options = {
-    expiresIn: "1d"
-  };
-
-  return jwt.sign(payload, secret, options);
-}
-
-router.post("/login", (req, res) => {
-  let { username, password } = req.body;
-
-  Users.findBy({ username })
-    .first()
-    .then(user => {
-      if (user && bcrypt.compareSync(password, user.password)) {
-        const token = generateToken(user);
-        res.status(200).json({ message: "this is a success message", token });
-      } else {
-        res.status(401).json({ message: "YOU SHALL NOT PASS!" });
-      }
-    })
-    .catch(error => {
-      res.status(500).json(error);
-    });
-});
-
-function restricted(req, res, next) {
-  const token = req.headers.authorization;
-
-  if (token) {
-    jwt.verify(token, secret, (error, decodedToken) => {
-      if (error) {
-        res.status(401).json({ message: "Can't touch this!" });
-      } else {
-        req.decodedJwt = decodedToken;
-        next();
-      }
-    });
-  } else {
-    res.status(401).json({ message: "YOU SHALL NOT PASS!" });
-  }
-}
-
-router.get("/users", restricted, (req, res) => {
-  console.log({ Users });
+router.get("/", restricted, (req, res) => {
   Users.find()
     .then(users => {
-      console.log({ users });
       res.status(200).json({ users, decodedToken: req.decodedJwt });
     })
     .catch(error => {
